*: make parser package dependency as small as possible (#7989)
