+++ ast/ast.go
@@ -19,7 +19,7 @@ import (
 	"io"
 
 	"github.com/pingcap/tidb/model"
-	"github.com/pingcap/tidb/types"
+	"github.com/pingcap/tidb/parser/types"
 )
 
 // Node is the basic element of the AST.
