+++ ast/base.go
@@ -13,7 +13,7 @@
 
 package ast
 
-import "github.com/pingcap/tidb/types"
+import "github.com/pingcap/tidb/parser/types"
 
 // node is the struct implements node interface except for Accept method.
 // Node implementations should embed it in.
