+++ ast/functions.go
@@ -18,7 +18,7 @@ import (
 	"io"
 
 	"github.com/pingcap/tidb/model"
-	"github.com/pingcap/tidb/types"
+	"github.com/pingcap/tidb/parser/types"
 )
 
 var (
