+++ types/errors.go
@@ -15,6 +15,7 @@ package types
 
 import (
 	"github.com/pingcap/tidb/mysql"
+	parser_types "github.com/pingcap/tidb/parser/types"
 	"github.com/pingcap/tidb/terror"
 )
 
@@ -45,12 +46,12 @@ var (
 	ErrWrongFieldSpec = terror.ClassTypes.New(codeWrongFieldSpec, "Wrong Field Spec")
 	// ErrBadNumber is return when parsing an invalid binary decimal number.
 	ErrBadNumber = terror.ClassTypes.New(codeBadNumber, "Bad Number")
+	// ErrInvalidDefault is returned when meet a invalid default value.
+	ErrInvalidDefault = parser_types.ErrInvalidDefault
 	// ErrCastAsSignedOverflow is returned when positive out-of-range integer, and convert to it's negative complement.
 	ErrCastAsSignedOverflow = terror.ClassTypes.New(codeUnknown, msgCastAsSignedOverflow)
 	// ErrCastNegIntAsUnsigned is returned when a negative integer be casted to an unsigned int.
 	ErrCastNegIntAsUnsigned = terror.ClassTypes.New(codeUnknown, msgCastNegIntAsUnsigned)
-	// ErrInvalidDefault is returned when meet a invalid default value.
-	ErrInvalidDefault = terror.ClassTypes.New(codeInvalidDefault, "Invalid default value for '%s'")
 	// ErrMBiggerThanD is returned when precision less than the scale.
 	ErrMBiggerThanD = terror.ClassTypes.New(codeMBiggerThanD, mysql.MySQLErrName[mysql.ErrMBiggerThanD])
 	// ErrWarnDataOutOfRange is returned when the value in a numeric column that is outside the permissible range of the column data type.
